Network.HTTP.Types.Header

Description

Type and constants for handling HTTP header fields.

At the bottom are also some functions to handle certain header field values.

HTTP Headers

type Header = (HeaderName, ByteString)

A full HTTP header field with the name and value separated.

E.g. "Content-Length: 28" parsed into a Header would turn into ("Content-Length", "28")

type HeaderName = CI ByteString

A case-insensitive name of a header field.

This is the part of the header field before the colon: HeaderName: some value

type RequestHeaders = [Header]

A list of Headers.

Same type as ResponseHeaders, but useful to differentiate in type signatures.

type ResponseHeaders = [Header]

A list of Headers.

Same type as RequestHeaders, but useful to differentiate in type signatures.

Common headers

The following header constants are provided for convenience, to prevent accidental spelling errors.

hAccept :: HeaderName

Accept

Since: 0.7.0

hAcceptCharset :: HeaderName

Accept-Charset

Since: 0.9

hAcceptEncoding :: HeaderName

Accept-Encoding

Since: 0.9

hAcceptLanguage :: HeaderName

Accept-Language

Since: 0.7.0

hAcceptRanges :: HeaderName

Accept-Ranges

Since: 0.9

hAge :: HeaderName

Age

Since: 0.9

hAllow :: HeaderName

Allow

Since: 0.9

hAuthorization :: HeaderName

Authorization

Since: 0.7.0

hCacheControl :: HeaderName

Cache-Control

Since: 0.7.0

hConnection :: HeaderName

Connection

Since: 0.7.0

hContentDisposition :: HeaderName

Content-Disposition

Since: 0.10

hContentEncoding :: HeaderName

Content-Encoding

Since: 0.7.0

hContentLanguage :: HeaderName

Content-Language

Since: 0.9

hContentLength :: HeaderName

Content-Length

Since: 0.7.0

hContentLocation :: HeaderName

Content-Location

Since: 0.9

hContentMD5 :: HeaderName

Content-MD5

This header has been obsoleted in RFC 9110.

Since: 0.7.0

hContentRange :: HeaderName

Content-Range

Since: 0.9

hContentType :: HeaderName

Content-Type

Since: 0.7.0

hCookie :: HeaderName

Cookie

Since: 0.7.0

hDate :: HeaderName

Date

Since: 0.7.0

hETag :: HeaderName

ETag

Since: 0.9

hExpect :: HeaderName

Expect

Since: 0.9

hExpires :: HeaderName

Expires

Since: 0.9

hFrom :: HeaderName

From

Since: 0.9

hHost :: HeaderName

Host

Since: 0.9

hIfMatch :: HeaderName

If-Match

Since: 0.9

hIfModifiedSince :: HeaderName

If-Modified-Since

Since: 0.7.0

hIfNoneMatch :: HeaderName

If-None-Match

Since: 0.9

hIfRange :: HeaderName

If-Range

Since: 0.7.0

hIfUnmodifiedSince :: HeaderName

If-Unmodified-Since

Since: 0.9

hLastModified :: HeaderName

Last-Modified

Since: 0.7.0

hLocation :: HeaderName

Location

Since: 0.7.1

hMaxForwards :: HeaderName

Max-Forwards

Since: 0.9

hMIMEVersion :: HeaderName

MIME-Version

Since: 0.10

hOrigin :: HeaderName

Origin

Since: 0.10

hPragma :: HeaderName

Pragma

This header has been deprecated in RFC 9111 in favor of "Cache-Control".

Since: 0.9

hPrefer :: HeaderName

Prefer

Since: 0.12.2

hPreferenceApplied :: HeaderName

Preference-Applied

Since: 0.12.2

hProxyAuthenticate :: HeaderName

Proxy-Authenticate

Since: 0.9

hProxyAuthorization :: HeaderName

Proxy-Authorization

Since: 0.9

hRange :: HeaderName

Range

Since: 0.7.0

hReferer :: HeaderName

Referer

Since: 0.7.0

hRetryAfter :: HeaderName

Retry-After

Since: 0.9

hServer :: HeaderName

Server

Since: 0.7.1

hSetCookie :: HeaderName

Set-Cookie

Since: 0.10

hTE :: HeaderName

TE

Since: 0.9

hTrailer :: HeaderName

Trailer

Since: 0.9

hTransferEncoding :: HeaderName

Transfer-Encoding

Since: 0.9

hUpgrade :: HeaderName

Upgrade

Since: 0.9

hUserAgent :: HeaderName

User-Agent

Since: 0.7.0

hVary :: HeaderName

Vary

Since: 0.9

hVia :: HeaderName

Via

Since: 0.9

hWWWAuthenticate :: HeaderName

WWW-Authenticate

Since: 0.9

hWarning :: HeaderName

Warning

This header has been obsoleted in RFC 9110.

Since: 0.9

Byte ranges

Convenience functions and types to handle values from Range headers.

https://www.rfc-editor.org/rfc/rfc9110.html#name-byte-ranges

data ByteRange

An individual byte range.

Negative indices are not allowed!

Since: 0.6.11

Constructors

ByteRangeFrom !Integer

ByteRangeFromTo !Integer !Integer

ByteRangeSuffix !Integer

renderByteRangeBuilder :: ByteRange -> Builder

Turns a byte range into a byte string Builder.

Since: 0.6.11

renderByteRange :: ByteRange -> ByteString

Renders a byte range into a ByteString.

>>> renderByteRange (ByteRangeFrom 2048)
"2048-"

Since: 0.6.11

type ByteRanges = [ByteRange]

A list of byte ranges.

Since: 0.6.11

renderByteRangesBuilder :: ByteRanges -> Builder

Turns a list of byte ranges into a byte string Builder.

Since: 0.6.11

renderByteRanges :: ByteRanges -> ByteString

Renders a list of byte ranges into a ByteString.

>>> renderByteRanges [ByteRangeFrom 2048, ByteRangeSuffix 20]
"bytes=2048-,-20"

Since: 0.6.11

parseByteRanges :: ByteString -> Maybe ByteRanges

Parse the value of a Range header into a ByteRanges.

>>> parseByteRanges "error"
Nothing
>>> parseByteRanges "bytes=0-499"
Just [ByteRangeFromTo 0 499]
>>> parseByteRanges "bytes=500-999"
Just [ByteRangeFromTo 500 999]
>>> parseByteRanges "bytes=-500"
Just [ByteRangeSuffix 500]
>>> parseByteRanges "bytes=9500-"
Just [ByteRangeFrom 9500]
>>> parseByteRanges "bytes=0-0,-1"
Just [ByteRangeFromTo 0 0,ByteRangeSuffix 1]
>>> parseByteRanges "bytes=500-600,601-999"
Just [ByteRangeFromTo 500 600,ByteRangeFromTo 601 999]
>>> parseByteRanges "bytes=500-700,601-999"
Just [ByteRangeFromTo 500 700,ByteRangeFromTo 601 999]

Since: 0.9.1