Network.HTTP.Types.Method

Description

Types and constants for HTTP methods.

The HTTP standard defines a set of standard methods, when to use them, and how to handle them. The standard set has been provided as a separate data type StdMethod, but since you can also use custom methods, the basic type Method is just a synonym for ByteString.

HTTP methods

type Method = ByteString

HTTP method (flat ByteString type).

Constants

methodGet :: Method

HTTP GET Method

methodPost :: Method

HTTP POST Method

methodHead :: Method

HTTP HEAD Method

methodPut :: Method

HTTP PUT Method

methodDelete :: Method

HTTP DELETE Method

methodTrace :: Method

HTTP TRACE Method

methodConnect :: Method

HTTP CONNECT Method

methodOptions :: Method

HTTP OPTIONS Method

methodPatch :: Method

HTTP PATCH Method

Since: 0.8.0

Standard Methods

One data type that holds all standard HTTP methods.

data StdMethod

HTTP standard method (as defined by RFC 2616, and PATCH which is defined by RFC 5789).

Since: 0.2.0

Constructors

GET

POST

HEAD

PUT

DELETE

TRACE

CONNECT

OPTIONS

PATCH
Since: 0.8.0

parseMethod :: Method -> Either ByteString StdMethod

Convert a method ByteString to a StdMethod if possible.

Since: 0.2.0

renderMethod :: Either ByteString StdMethod -> Method

Convert an algebraic method to a ByteString.

renderMethod (parseMethod bs) == bs

Since: 0.3.0

renderStdMethod :: StdMethod -> Method

Convert a StdMethod to a ByteString.

Since: 0.2.0