Network.HTTP.Types

Methods

For more information: Network.HTTP.Types.Method

type Method = ByteString

HTTP method (flat ByteString type).

Constants

methodGet :: Method

HTTP GET Method

methodPost :: Method

HTTP POST Method

methodHead :: Method

HTTP HEAD Method

methodPut :: Method

HTTP PUT Method

methodDelete :: Method

HTTP DELETE Method

methodTrace :: Method

HTTP TRACE Method

methodConnect :: Method

HTTP CONNECT Method

methodOptions :: Method

HTTP OPTIONS Method

methodPatch :: Method

HTTP PATCH Method

Since: 0.8.0

data StdMethod

HTTP standard method (as defined by RFC 2616, and PATCH which is defined by RFC 5789).

Since: 0.2.0

Constructors

GET

POST

HEAD

PUT

DELETE

TRACE

CONNECT

OPTIONS

PATCH
Since: 0.8.0

Parsing and redering methods

parseMethod :: Method -> Either ByteString StdMethod

Convert a method ByteString to a StdMethod if possible.

Since: 0.2.0

renderMethod :: Either ByteString StdMethod -> Method

Convert an algebraic method to a ByteString.

renderMethod (parseMethod bs) == bs

Since: 0.3.0

renderStdMethod :: StdMethod -> Method

Convert a StdMethod to a ByteString.

Since: 0.2.0

Versions

For more information: Network.HTTP.Types.Version

data HttpVersion

HTTP Version.

Note that the Show instance is intended merely for debugging.

Constructors

HttpVersion

Fields

httpMajor :: !Int

httpMinor :: !Int

http09 :: HttpVersion

HTTP 0.9

http10 :: HttpVersion

HTTP 1.0

http11 :: HttpVersion

HTTP 1.1

http20 :: HttpVersion

HTTP 2.0

Since: 0.10

Status

For more information: Network.HTTP.Types.Status

data Status

HTTP Status.

Only the statusCode is used for comparisons.

Please use mkStatus to create status codes from code and message, or the Enum instance or the status code constants (like ok200). There might be additional record members in the future.

Note that the Show instance is only for debugging.

Constructors

Status

Fields

statusCode :: Int

statusMessage :: ByteString

Constants

mkStatus :: Int -> ByteString -> Status

Create a Status from a status code and message.

status100 :: Status

Continue 100

Since: 0.6.6

continue100 :: Status

Continue 100

Since: 0.6.6

status101 :: Status

Switching Protocols 101

Since: 0.6.6

switchingProtocols101 :: Status

Switching Protocols 101

Since: 0.6.6

status200 :: Status

OK 200

ok200 :: Status

OK 200

status201 :: Status

Created 201

created201 :: Status

Created 201

status202 :: Status

Accepted 202

Since: 0.6.6

accepted202 :: Status

Accepted 202

Since: 0.6.6

status203 :: Status

Non-Authoritative Information 203

Since: 0.6.6

nonAuthoritative203 :: Status

Non-Authoritative Information 203

Since: 0.6.6

status204 :: Status

No Content 204

Since: 0.6.6

noContent204 :: Status

No Content 204

Since: 0.6.6

status205 :: Status

Reset Content 205

Since: 0.6.6

resetContent205 :: Status

Reset Content 205

Since: 0.6.6

status206 :: Status

Partial Content 206

Since: 0.5.1

partialContent206 :: Status

Partial Content 206

Since: 0.5.1

status300 :: Status

Multiple Choices 300

multipleChoices300 :: Status

Multiple Choices 300

status301 :: Status

Moved Permanently 301

movedPermanently301 :: Status

Moved Permanently 301

status302 :: Status

Found 302

found302 :: Status

Found 302

status303 :: Status

See Other 303

seeOther303 :: Status

See Other 303

status304 :: Status

Not Modified 304

Since: 0.6.1

notModified304 :: Status

Not Modified 304

Since: 0.6.1

status305 :: Status

Use Proxy 305

Since: 0.6.6

useProxy305 :: Status

Use Proxy 305

Since: 0.6.6

status307 :: Status

Temporary Redirect 307

Since: 0.6.6

temporaryRedirect307 :: Status

Temporary Redirect 307

Since: 0.6.6

status308 :: Status

Permanent Redirect 308

Since: 0.9

permanentRedirect308 :: Status

Permanent Redirect 308

Since: 0.9

status400 :: Status

Bad Request 400

badRequest400 :: Status

Bad Request 400

status401 :: Status

Unauthorized 401

unauthorized401 :: Status

Unauthorized 401

status402 :: Status

Payment Required 402

Since: 0.6.6

paymentRequired402 :: Status

Payment Required 402

Since: 0.6.6

status403 :: Status

Forbidden 403

forbidden403 :: Status

Forbidden 403

status404 :: Status

Not Found 404

notFound404 :: Status

Not Found 404

status405 :: Status

Method Not Allowed 405

methodNotAllowed405 :: Status

Method Not Allowed 405

status406 :: Status

Not Acceptable 406

Since: 0.6.6

notAcceptable406 :: Status

Not Acceptable 406

Since: 0.6.6

status407 :: Status

Proxy Authentication Required 407

Since: 0.6.6

proxyAuthenticationRequired407 :: Status

Proxy Authentication Required 407

Since: 0.6.6

status408 :: Status

Request Timeout 408

Since: 0.6.6

requestTimeout408 :: Status

Request Timeout 408

Since: 0.6.6

status409 :: Status

Conflict 409

Since: 0.6.6

conflict409 :: Status

Conflict 409

Since: 0.6.6

status410 :: Status

Gone 410

Since: 0.6.6

gone410 :: Status

Gone 410

Since: 0.6.6

status411 :: Status

Length Required 411

Since: 0.6.6

lengthRequired411 :: Status

Length Required 411

Since: 0.6.6

status412 :: Status

Precondition Failed 412

Since: 0.6.1

preconditionFailed412 :: Status

Precondition Failed 412

Since: 0.6.1

status413 :: Status

Request Entity Too Large 413

Since: 0.6.6

requestEntityTooLarge413 :: Status

Request Entity Too Large 413

Since: 0.6.6

status414 :: Status

Request-URI Too Long 414

Since: 0.6.6

requestURITooLong414 :: Status

Request-URI Too Long 414

Since: 0.6.6

status415 :: Status

Unsupported Media Type 415

Since: 0.6.6

unsupportedMediaType415 :: Status

Unsupported Media Type 415

Since: 0.6.6

status416 :: Status

Requested Range Not Satisfiable 416

Since: 0.6.1

requestedRangeNotSatisfiable416 :: Status

Requested Range Not Satisfiable 416

Since: 0.6.1

status417 :: Status

Expectation Failed 417

Since: 0.6.6

expectationFailed417 :: Status

Expectation Failed 417

Since: 0.6.6

status418 :: Status

I'm a teapot 418

Since: 0.6.6

imATeapot418 :: Status

I'm a teapot 418

Since: 0.6.6

status422 :: Status

Unprocessable Entity 422 (RFC 4918)

Since: 0.9.1

unprocessableEntity422 :: Status

Unprocessable Entity 422 (RFC 4918)

Since: 0.9.1

status428 :: Status

Precondition Required 428 (RFC 6585)

Since: 0.8.5

preconditionRequired428 :: Status

Precondition Required 428 (RFC 6585)

Since: 0.8.5

status429 :: Status

Too Many Requests 429 (RFC 6585)

Since: 0.8.5

tooManyRequests429 :: Status

Too Many Requests 429 (RFC 6585)

Since: 0.8.5

status431 :: Status

Request Header Fields Too Large 431 (RFC 6585)

Since: 0.8.5

requestHeaderFieldsTooLarge431 :: Status

Request Header Fields Too Large 431 (RFC 6585)

Since: 0.8.5

status500 :: Status

Internal Server Error 500

internalServerError500 :: Status

Internal Server Error 500

status501 :: Status

Not Implemented 501

Since: 0.6.1

notImplemented501 :: Status

Not Implemented 501

Since: 0.6.1

status502 :: Status

Bad Gateway 502

Since: 0.6.6

badGateway502 :: Status

Bad Gateway 502

Since: 0.6.6

status503 :: Status

Service Unavailable 503

Since: 0.6.6

serviceUnavailable503 :: Status

Service Unavailable 503

Since: 0.6.6

status504 :: Status

Gateway Timeout 504

Since: 0.6.6

gatewayTimeout504 :: Status

Gateway Timeout 504

Since: 0.6.6

status505 :: Status

HTTP Version Not Supported 505

Since: 0.6.6

httpVersionNotSupported505 :: Status

HTTP Version Not Supported 505

Since: 0.6.6

status511 :: Status

Network Authentication Required 511 (RFC 6585)

Since: 0.8.5

networkAuthenticationRequired511 :: Status

Network Authentication Required 511 (RFC 6585)

Since: 0.8.5

statusIsInformational :: Status -> Bool

Informational class

Checks if the status is in the 1XX range.

Since: 0.8.0

statusIsSuccessful :: Status -> Bool

Successful class

Checks if the status is in the 2XX range.

Since: 0.8.0

statusIsRedirection :: Status -> Bool

Redirection class

Checks if the status is in the 3XX range.

Since: 0.8.0

statusIsClientError :: Status -> Bool

Client Error class

Checks if the status is in the 4XX range.

Since: 0.8.0

statusIsServerError :: Status -> Bool

Server Error class

Checks if the status is in the 5XX range.

Since: 0.8.0

Headers

For more information: Network.HTTP.Types.Header

Types

type Header = (HeaderName, ByteString)

A full HTTP header field with the name and value separated.

E.g. "Content-Length: 28" parsed into a Header would turn into ("Content-Length", "28")

type HeaderName = CI ByteString

A case-insensitive name of a header field.

This is the part of the header field before the colon: HeaderName: some value

type RequestHeaders = [Header]

A list of Headers.

Same type as ResponseHeaders, but useful to differentiate in type signatures.

type ResponseHeaders = [Header]

A list of Headers.

Same type as RequestHeaders, but useful to differentiate in type signatures.

Common headers

hAccept :: HeaderName

Accept

Since: 0.7.0

hAcceptLanguage :: HeaderName

Accept-Language

Since: 0.7.0

hAuthorization :: HeaderName

Authorization

Since: 0.7.0

hCacheControl :: HeaderName

Cache-Control

Since: 0.7.0

hCookie :: HeaderName

Cookie

Since: 0.7.0

hConnection :: HeaderName

Connection

Since: 0.7.0

hContentEncoding :: HeaderName

Content-Encoding

Since: 0.7.0

hContentLength :: HeaderName

Content-Length

Since: 0.7.0

hContentMD5 :: HeaderName

Content-MD5

This header has been obsoleted in RFC 9110.

Since: 0.7.0

hContentType :: HeaderName

Content-Type

Since: 0.7.0

hDate :: HeaderName

Date

Since: 0.7.0

hIfModifiedSince :: HeaderName

If-Modified-Since

Since: 0.7.0

hIfRange :: HeaderName

If-Range

Since: 0.7.0

hLastModified :: HeaderName

Last-Modified

Since: 0.7.0

hLocation :: HeaderName

Location

Since: 0.7.1

hRange :: HeaderName

Range

Since: 0.7.0

hReferer :: HeaderName

Referer

Since: 0.7.0

hServer :: HeaderName

Server

Since: 0.7.1

hUserAgent :: HeaderName

User-Agent

Since: 0.7.0

Byte ranges

data ByteRange

An individual byte range.

Negative indices are not allowed!

Since: 0.6.11

Constructors

ByteRangeFrom !Integer

ByteRangeFromTo !Integer !Integer

ByteRangeSuffix !Integer

renderByteRangeBuilder :: ByteRange -> Builder

Turns a byte range into a byte string Builder.

Since: 0.6.11

renderByteRange :: ByteRange -> ByteString

Renders a byte range into a ByteString.

>>> renderByteRange (ByteRangeFrom 2048)
"2048-"

Since: 0.6.11

type ByteRanges = [ByteRange]

A list of byte ranges.

Since: 0.6.11

renderByteRangesBuilder :: ByteRanges -> Builder

Turns a list of byte ranges into a byte string Builder.

Since: 0.6.11

renderByteRanges :: ByteRanges -> ByteString

Renders a list of byte ranges into a ByteString.

>>> renderByteRanges [ByteRangeFrom 2048, ByteRangeSuffix 20]
"bytes=2048-,-20"

Since: 0.6.11

parseByteRanges :: ByteString -> Maybe ByteRanges

Parse the value of a Range header into a ByteRanges.

>>> parseByteRanges "error"
Nothing
>>> parseByteRanges "bytes=0-499"
Just [ByteRangeFromTo 0 499]
>>> parseByteRanges "bytes=500-999"
Just [ByteRangeFromTo 500 999]
>>> parseByteRanges "bytes=-500"
Just [ByteRangeSuffix 500]
>>> parseByteRanges "bytes=9500-"
Just [ByteRangeFrom 9500]
>>> parseByteRanges "bytes=0-0,-1"
Just [ByteRangeFromTo 0 0,ByteRangeSuffix 1]
>>> parseByteRanges "bytes=500-600,601-999"
Just [ByteRangeFromTo 500 600,ByteRangeFromTo 601 999]
>>> parseByteRanges "bytes=500-700,601-999"
Just [ByteRangeFromTo 500 700,ByteRangeFromTo 601 999]

Since: 0.9.1

URI

For more extensive information: Network.HTTP.Types.URI

Query strings

Query

type Query = [QueryItem]

A sequence of QueryItems.

type QueryItem = (ByteString, Maybe ByteString)

An item from the query string, split up into two parts.

The second part should be Nothing if there was no key-value separator after the query item name.

Since: 0.2.0

renderQuery :: Bool -> Query -> ByteString

Renders the given Query into a ByteString.

If you want a question mark (?) added to the front of the result, use True.

Since: 0.2.0

renderQueryBuilder :: Bool -> Query -> Builder

Renders the given Query into a Builder.

If you want a question mark (?) added to the front of the result, use True.

Since: 0.5

parseQuery :: ByteString -> Query

Split out the query string into a list of keys and values. A few importants points:

The result returned is still bytestrings, since we perform no character decoding here. Most likely, you will want to use UTF-8 decoding, but this is left to the user of the library.

Percent decoding errors are ignored. In particular, "%Q" will be output as "%Q".

It decodes '+' characters to ' '

Since: 0.2.0

parseQueryReplacePlus :: Bool -> ByteString -> Query

Same functionality as parseQuery, but with the option to decode '+' characters to ' ' or to preserve any '+' encountered.

If you want to replace any '+' with a space, use True.

Since: 0.12.2

Text query string (UTF8 encoded)

type QueryText = [(Text, Maybe Text)]

Like Query, but with Text instead of ByteString (UTF8-encoded).

Since: 0.5.2

queryTextToQuery :: QueryText -> Query

Convert QueryText to Query.

Since: 0.5.2

queryToQueryText :: Query -> QueryText

Convert Query to QueryText (leniently decoding the UTF-8).

Since: 0.5.2

renderQueryText :: Bool -> QueryText -> Builder

Convert QueryText to a Builder.

If you want a question mark (?) added to the front of the result, use True.

Since: 0.5.2

parseQueryText :: ByteString -> QueryText

Parse a QueryText from a ByteString. See parseQuery for details.

queryToQueryText . parseQuery

Since: 0.5.2

SimpleQuery

type SimpleQuery = [SimpleQueryItem]

A sequence of SimpleQueryItems.

type SimpleQueryItem = (ByteString, ByteString)

Simplified query item type without support for parameter-less items.

Since: 0.2.0

simpleQueryToQuery :: SimpleQuery -> Query

Convert SimpleQuery to Query.

Since: 0.5

renderSimpleQuery :: Bool -> SimpleQuery -> ByteString

Render the given SimpleQuery into a ByteString.

If you want a question mark (?) added to the front of the result, use True.

Since: 0.2.0

parseSimpleQuery :: ByteString -> SimpleQuery

Parse SimpleQuery from a ByteString.

This uses parseQuery under the hood, and will transform any Nothing values into an empty ByteString.

Since: 0.2.0

PartialEscapeQuery

type PartialEscapeQuery = [PartialEscapeQueryItem]

Query with some characters that should not be escaped.

General form: a=b&c=d:e+f&g=h

Since: 0.12.1

type PartialEscapeQueryItem = (ByteString, [EscapeItem])

Partially escaped query item.

The key will always be encoded using 'urlEncode True', but the value will be encoded depending on which EscapeItems are used.

Since: 0.12.1

data EscapeItem

Section of a query item value that decides whether to use regular URL encoding (using 'urlEncode True') with QE, or to not encode anything with QN.

Since: 0.12.1

Constructors

QE ByteString
will be URL encoded

QN ByteString
will NOT at all be URL encoded

renderQueryPartialEscape :: Bool -> PartialEscapeQuery -> ByteString

Convert PartialEscapeQuery to ByteString.

If you want a question mark (?) added to the front of the result, use True.

>>> renderQueryPartialEscape True [("a", [QN "x:z + ", QE (encodeUtf8 "They said: \"שלום\"")])]
"?a=x:z + They%20said%3A%20%22%D7%A9%D7%9C%D7%95%D7%9D%22"

Since: 0.12.1

renderQueryBuilderPartialEscape :: Bool -> PartialEscapeQuery -> Builder

Convert a PartialEscapeQuery to a Builder.

If you want a question mark (?) added to the front of the result, use True.

Since: 0.12.1

Generalized query types

class QueryLike a where

Types which can, and commonly are, converted to Query are in this class.

You can use lists of simple key value pairs, with ByteString (strict, or lazy: ByteString), Text, or String as the key/value types. You can also have the value type lifted into a Maybe to support keys without values; and finally it is possible to put each pair into a Maybe for key-value pairs that aren't always present.

Since: 0.7.0

Methods

toQuery :: a -> Query

Convert to Query.

Path

Segments + Query String

extractPath :: ByteString -> ByteString

Extract whole path (path segments + query) from a RFC 2616 Request-URI.

Though a more accurate description of this function's behaviour is that it removes the domain/origin if the string starts with an HTTP protocol. (i.e. http:// or https://)

This function will not change anything when given any other ByteString. (except return a root path "/" if given an empty string)

>>> extractPath "/path"
"/path"

>>> extractPath "http://example.com:8080/path"
"/path"

>>> extractPath "http://example.com"
"/"

>>> extractPath ""
"/"

>>> extractPath "www.google.com/some/path"
"www.google.com/some/path"

Since: 0.8.5

encodePath :: [Text] -> Query -> Builder

Encode a whole path (path segments + query).

Since: 0.5

decodePath :: ByteString -> ([Text], Query)

Decode a whole path (path segments + query).

Since: 0.5

Path Segments

encodePathSegments :: [Text] -> Builder

Encodes a list of path segments into a valid URL fragment.

This function takes the following three steps:

UTF-8 encodes the characters.

Prepends each segment with a slash.

Performs percent-encoding on all characters that are not:

alphanumeric (i.e. A-Z and a-z)

digits (i.e. 0-9)

a dash '-', an underscore '_', a dot '.', or a tilde '~'

For example:

>>> encodePathSegments ["foo", "bar1", "~baz"]
"/foo/bar1/~baz"

>>> encodePathSegments ["foo bar", "baz/bin"]
"/foo%20bar/baz%2Fbin"

>>> encodePathSegments ["שלום"]
"/%D7%A9%D7%9C%D7%95%D7%9D"

Huge thanks to Jeremy Shaw who created the original implementation of this function in web-routes and did such thorough research to determine all correct escaping procedures.

Since: 0.5

encodePathSegmentsRelative :: [Text] -> Builder

Like encodePathSegments, but without the initial slash.

Since: 0.6.10

decodePathSegments :: ByteString -> [Text]

Parse a list of path segments from a valid URL fragment.

Will also decode any percent-encoded characters.

Since: 0.5

URL encoding / decoding

urlEncode :: Bool -> ByteString -> ByteString

Percent-encoding for URLs.

In short:

if you're encoding (parts of) a path element, use False.

if you're encoding (parts of) a query string, use True.

In-depth explanation

Expand

This will substitute every byte with its percent-encoded equivalent unless:

The byte is alphanumeric. (i.e. A-Z, a-z, or 0-9)

The byte is either a dash '-', an underscore '_', a dot '.', or a tilde '~'

If False is used, the following will also not be percent-encoded:

colon ':', at sign '@', ampersand '&', equals sign '=', plus sign '+', dollar sign '$' or a comma ','

Since: 0.2.0

urlEncodeBuilder :: Bool -> ByteString -> Builder

Percent-encoding for URLs.

Like urlEncode, but only makes the Builder.

Since: 0.5

urlDecode :: Bool -> ByteString -> ByteString

Percent-decoding.

If you want to replace any '+' with a space, use True.

Since: 0.2.0